% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rins-package.R
\docType{package}
\name{rins-package}
\alias{rins}
\alias{rins-package}
\title{rins: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
